combine data and policy trait impl, consolidate in a single file
