+++ crates/bitwarden-policies/src/lib.rs
@@ -7,19 +7,15 @@ mod uniffi_support;
 
 mod enriched_policy_type;
 pub mod filter;
-mod master_password_policy_response;
+pub mod policy_definitions;
 mod models;
-pub mod policy_data;
 mod policy_client;
-pub mod policy_overrides;
 mod policy_type;
 mod registry;
 
 pub use enriched_policy_type::EnrichedPolicyType;
 pub use filter::Policy;
-pub use master_password_policy_response::MasterPasswordPolicyResponse;
 pub use models::{OrganizationUserPolicyContext, PolicyView};
 pub use policy_client::{PoliciesClientExt, PolicyClient};
-pub use policy_data::*;
-pub use policy_overrides::*;
+pub use policy_definitions::*;
 pub use policy_type::PolicyType;
